feat: create routers folder and group path operations.
